+++ MHGoldenJuteApp/MHGoldenJuteApp/Models/EntityModel/Item.cs
@@ -7,8 +7,8 @@ namespace MHGoldenJuteApp.Models
 {
     public class Item
     {
-        public int ItemId { get; set; }
-        public int ItemGroupId { get; set; }
+        public Int64 ItemId { get; set; }
+        public Int64 ItemGroupId { get; set; }
         public string ItemCode { get; set; }
         public string ItemName { get; set; }
         public string Department { get; set; }
